+++ Infrastructure/Repository/IRepositoryUser.cs
@@ -11,7 +11,7 @@ namespace Infrastructure.Repository
     {
         IEnumerable<User> GetUsers();
         User GetUserByID(int id);
-        User GetUserByIDForLogin(int id);
+        User GetUserByIDForLogin(long id);
         User Save(User user);
         User GetUsersForLogin(string email, string password);
     }
